+++ rusk/src/bin/config/chain.rs
@@ -22,18 +22,22 @@ pub(crate) struct ChainConfig {
     consensus_keys_path: Option<PathBuf>,
     #[serde(with = "humantime_serde")]
     #[serde(default)]
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     generation_timeout: Option<Duration>,
 
     max_queue_size: Option<usize>,
 
     // NB: changing the gas_per_deploy_byte/block_gas_limit is equivalent to
     // forking the chain.
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     gas_per_deploy_byte: Option<u64>,
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     min_deployment_gas_price: Option<u64>,
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     min_deploy_points: Option<u64>,
     min_gas_limit: Option<u64>,
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     block_gas_limit: Option<u64>,
-    public_sender_start_height: Option<u64>,
 
     #[serde(with = "humantime_serde")]
     #[serde(default)]
@@ -81,24 +85,28 @@ impl ChainConfig {
         self.db_options.clone().unwrap_or_default()
     }
 
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     pub(crate) fn generation_timeout(&self) -> Option<Duration> {
+        #[allow(deprecated)]
         self.generation_timeout
     }
 
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     pub(crate) fn gas_per_deploy_byte(&self) -> Option<u64> {
+        #[allow(deprecated)]
         self.gas_per_deploy_byte
     }
 
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     pub(crate) fn min_deployment_gas_price(&self) -> Option<u64> {
+        #[allow(deprecated)]
         self.min_deployment_gas_price
     }
 
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     pub(crate) fn min_deploy_points(&self) -> Option<u64> {
+        #[allow(deprecated)]
         self.min_deploy_points
-    }
-
-    pub(crate) fn public_sender_start_height(&self) -> Option<u64> {
-        self.public_sender_start_height
     }
 
     pub(crate) fn min_gas_limit(&self) -> Option<u64> {
@@ -109,7 +117,9 @@ impl ChainConfig {
         self.max_queue_size.unwrap_or(10_000)
     }
 
+    #[deprecated(since = "1.0.3", note = "please use `RuskVmConfig` instead")]
     pub(crate) fn block_gas_limit(&self) -> Option<u64> {
+        #[allow(deprecated)]
         self.block_gas_limit
     }