+++ rusk/src/lib/node/vm/config.rs
@@ -4,32 +4,52 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use std::collections::HashMap;
 use std::time::Duration;
 
 use dusk_vm::ExecutionConfig;
+use serde::{Deserialize, Serialize};
 
-const DEFAULT_GAS_PER_DEPLOY_BYTE: u64 = 100;
-const DEFAULT_MIN_DEPLOYMENT_GAS_PRICE: u64 = 2000;
-const DEFAULT_MIN_DEPLOY_POINTS: u64 = 5_000_000;
-const DEFAULT_BLOCK_GAS_LIMIT: u64 = 3 * 1_000_000_000;
+const fn default_gas_per_deploy_byte() -> u64 {
+    100
+}
+const fn default_min_deploy_points() -> u64 {
+    5_000_000
+}
+const fn default_min_deployment_gas_price() -> u64 {
+    2_000
+}
+const fn default_block_gas_limit() -> u64 {
+    3 * 1_000_000_000
+}
 
 /// Configuration for the execution of a transaction.
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Config {
     /// The amount of gas points charged for each byte in a contract-deployment
     /// bytecode.
+    #[serde(default = "default_gas_per_deploy_byte")]
     pub gas_per_deploy_byte: u64,
+
     /// The minimum gas points charged for a contract deployment.
+    #[serde(default = "default_min_deploy_points")]
     pub min_deploy_points: u64,
+
     /// The minimum gas price set for a contract deployment
-    pub min_deploy_gas_price: u64,
+    #[serde(default = "default_min_deployment_gas_price")]
+    pub min_deployment_gas_price: u64,
+
     /// The maximum amount of gas points that can be used in a block.
+    #[serde(default = "default_block_gas_limit")]
     pub block_gas_limit: u64,
+
     /// The timeout for a candidate block generation.
+    #[serde(with = "humantime_serde")]
+    #[serde(default)]
     pub generation_timeout: Option<Duration>,
-    /// The height at which the public sender starts to be injected in the
-    /// block metadata
-    pub public_sender_start_height: u64,
+
+    /// Set of features to activate
+    pub features: HashMap<String, u64>,
 }
 
 impl Default for Config {
@@ -41,13 +61,12 @@ impl Default for Config {
 impl Config {
     pub fn new() -> Self {
         Self {
-            gas_per_deploy_byte: DEFAULT_GAS_PER_DEPLOY_BYTE,
-            min_deploy_gas_price: DEFAULT_MIN_DEPLOYMENT_GAS_PRICE,
-            min_deploy_points: DEFAULT_MIN_DEPLOY_POINTS,
-            block_gas_limit: DEFAULT_BLOCK_GAS_LIMIT,
+            gas_per_deploy_byte: default_gas_per_deploy_byte(),
+            min_deployment_gas_price: default_min_deployment_gas_price(),
+            min_deploy_points: default_min_deploy_points(),
+            block_gas_limit: default_block_gas_limit(),
             generation_timeout: None,
-            // Disabled by default
-            public_sender_start_height: u64::MAX,
+            features: HashMap::new(),
         }
     }
 
@@ -81,7 +100,7 @@ impl Config {
         mut self,
         min_deploy_gas_price: u64,
     ) -> Self {
-        self.min_deploy_gas_price = min_deploy_gas_price;
+        self.min_deployment_gas_price = min_deploy_gas_price;
         self
     }
 
@@ -96,13 +115,22 @@ impl Config {
 
     /// Create a new `Config` with the given parameters.
     pub fn to_execution_config(&self, block_height: u64) -> ExecutionConfig {
-        let with_public_sender =
-            block_height >= self.public_sender_start_height;
+        let with_public_sender = self
+            .feature("ABI_PUBLIC_SENDER")
+            .map(|activation| activation >= block_height)
+            .unwrap_or_default();
         ExecutionConfig {
             gas_per_deploy_byte: self.gas_per_deploy_byte,
             min_deploy_points: self.min_deploy_points,
-            min_deploy_gas_price: self.min_deploy_gas_price,
+            min_deploy_gas_price: self.min_deployment_gas_price,
             with_public_sender,
         }
+    }
+
+    pub fn feature(&self, feature: &str) -> Option<u64> {
+        self.features
+            .iter()
+            .find(|(k, _)| k.eq_ignore_ascii_case(feature))
+            .map(|(_, &v)| v)
     }
 }